Streamlining Burst Detection
Extended per-topic burst detection to use mapWithState() and reduced the
size of the tweet window to the last minor-window size. Using
mapWithState() on tokenized sliding windows reduces complexity and lets
Spark handle the major window sizing I was doing manually in past
versions.

The new version seems significantly faster and is able to run in real
time for four topics.