Added Config Option for Replay and Debugging
Prior version used a hard-coded flag to enable/disable replay-mode,
which uses a socket stream to pull in data rather than the Twitter
stream. Since we've added support for Kafka as well, we now add a
configuration option to specify whether you're using Kafka, replay, or
defaults to Twitter's stream.